Hello Guys! Commit Update
I've added a view for the alert and a way to access it and return to
HomeView. The files effected: Added AlertView, Added
AlertController made changes to HomeView to support AlertView. This is
only a display at the moment so I'll be working on the logic and should
have updates by the end of tommorrow.