Buy code running
After the new set up had to pass the JFrame in order to function the
keys. I know some of you may think thats not a good idea. if you can fix
it.

every thing is running smoothly,but the regexpression function detect
errors in digits such as more than 99 but it does not react to letters
and I will leave this work for later.

please try the sell as soon as the Sell button works.